chisel_ng/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Jinja2==3.1.2
python-crontab==2.7.1